don't attempt to `define` `locale_t`
This has been creating issues on some toolchains, namely the latest
XCode toolchains. Having said this, we notice that `locale_t` is not
actually needed *anywhere*. It's used as an argument in `_strptime()`,
but the value of this is always 0 and it's effectively ignored. I think
th use of this originated from the original FreeBSD origin of this code
and then all the locale-specific stuff was stripped away by `define`-ing
`locale_t` as an `int` and just passing the value as `0` everywhere. In
this commit we just do what they should have done in the first place and
simply remove the argument.

We remove the `locale_t` argument from `_strptime()` and hard-code
`locale` to be the integer 0 in all other cases, which it was in the
first place.